fix: repair get_opponent_match_history, which raised on every call (B-39)
CI caught this: the smoke test added for B-38 failed with "structure of
query does not match function result type".

The cause is older than that change. The function unions three sources.
matches.team1_score and matches_archive.team1_score are integer, but
playoff_matches.team1_score and team2_score are numeric - that split is
in the baseline schema itself. UNION ALL resolves those output columns
to numeric, which does not match the integer the function declares, so
every call raised. Type resolution happens when the statement is
planned, so it failed whether or not a playoff match existed.

Reproduced locally by replaying every migration into a fresh Postgres:
the pre-existing definition raises the same error on a bare call. So
the head-to-head details dialog has been dead since 20250906000458 -
HeadToHeadService.getOpponentHistory throws, the query has no data, and
OpponentHistoryModal returns null on !history?.summary. Pressing an
opponent row did nothing at all, with no error surfaced. Filed as B-39.

The playoff branch now casts its scores to integer, matching the other
two sources and the declared return type. Scores are whole numbers
everywhere else in the schema.

The smoke test now also seeds a playoff match and pins its scores and
winner_id, so the branch that was broken carries data rather than only
contributing a plan.

Verified by replaying the bootstrap and all migrations into a fresh
Postgres and running all 20 smoke tests: green.

+++ supabase/migrations/20260826190000_opponent_match_history_winner_id.sql
@@ -17,6 +17,13 @@
 --
 -- RETURNS TABLE changes shape, so the function must be dropped first; CREATE OR
 -- REPLACE cannot change a return type.
+--
+-- This also repairs a fault the function has carried since it was created in
+-- 20250906000458: playoff_matches.team1_score/team2_score are numeric while the
+-- matches and matches_archive columns are integer, so UNION ALL resolved those
+-- output columns to numeric and every call raised "structure of query does not
+-- match function result type". The dialog that calls it therefore rendered
+-- nothing at all. The playoff branch now casts to integer.
 
 DROP FUNCTION IF EXISTS public.get_opponent_match_history(uuid, uuid);
 
@@ -119,11 +126,17 @@ BEGIN
       pm.team2_id,
       t1.name as team1_name,
       t2.name as team2_name,
-      pm.team1_score,
-      pm.team2_score,
+      -- playoff_matches.team1_score/team2_score are numeric while the same
+      -- columns on matches and matches_archive are integer. UNION ALL resolves
+      -- the branch types to numeric, which does not match the integer declared
+      -- below, so every call raised "structure of query does not match function
+      -- result type" before these casts. Scores are whole numbers everywhere
+      -- else in the schema.
+      pm.team1_score::integer,
+      pm.team2_score::integer,
       -- For playoff matches, calculate game wins from the scores
-      CASE WHEN pm.team1_score IS NOT NULL THEN pm.team1_score ELSE 0 END as team1_game_wins,
-      CASE WHEN pm.team2_score IS NOT NULL THEN pm.team2_score ELSE 0 END as team2_game_wins,
+      CASE WHEN pm.team1_score IS NOT NULL THEN pm.team1_score ELSE 0 END::integer as team1_game_wins,
+      CASE WHEN pm.team2_score IS NOT NULL THEN pm.team2_score ELSE 0 END::integer as team2_game_wins,
       CASE
         WHEN pm.winner_id = pm.team1_id THEN pm.team1_id
         WHEN pm.winner_id = pm.team2_id THEN pm.team2_id
